pyproject.toml
[tool.poetry]
name = "qRacee"
version = "0.1.0"
description = "qBittorrent Race Re-announcer"
authors = ["Gunnar Cortes Heimisson <gnarr.c@gmail.com>"]
license = "The Unlicense"

[tool.poetry.dependencies]
python = "^3.10"
qbittorrent-api = "^2022.5.32"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
